# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Identity Onboarding Lab — Public API
  version: 1.0.0
  description: Public HTTP API contract for Sri Lankan investment customer onboarding and reviewer staff operations.
paths:
...
  /applications:
    post:
      summary: Create resumable onboarding draft
      operationId: createApplication
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      responses:
        '201':
          description: Application draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
  /applications/{id}:
    get:
      summary: Fetch application state and next action
      operationId: getApplication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
...
  /staff/applications:
    get:
      summary: Reviewer queue of applications
      operationId: getStaffQueue
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of applications for review
  /staff/applications/{id}:
    get:
      summary: Reviewer detail view with masked PII and extracted differences
      operationId: getStaffApplicationDetail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review case detail
  /staff/applications/{id}/decision:
    post:
      summary: Reviewer records manual decision with mandatory reason
      operationId: recordStaffDecision
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion, decision, reason]
              properties:
                expectedVersion:
                  type: integer
                decision:
                  type: string
                  enum: [APPROVED, REJECTED, RECAPTURE_REQUIRED]
                reason:
                  type: string
      responses:
        '200':
          description: Decision updated
  /staff/evidence/{evidenceId}:
    get:
      summary: Authorized streaming of protected evidence with access audit
      operationId: streamEvidence
      parameters:
        - in: path
          name: evidenceId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Binary media stream
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
            video/webm:
              schema:
                type: string
                format: binary
...
components:
  schemas:
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lifecycle:
          type: string
          enum: [DRAFT, PROCESSING, READY_TO_SUBMIT, SUBMITTED, UNDER_REVIEW, COMPLETED, WITHDRAWN, REJECTED]
        decision:
          type: string
          enum: [PENDING, RECAPTURE_REQUIRED, REVIEW_REQUIRED, DEMO_CHECKS_PASSED, APPROVED, REJECTED]
        operatingMode:
          type: string
          enum: [LOCAL_DEMO, INSTITUTION]
        version:
          type: integer
        currentDocumentGeneration:
          type: integer
        currentBiometricGeneration:
          type: integer
        submittedAt:
          type: string
          format: date-time
          nullable: true
...